requirements.txt
discord.py
asyncpg
pinecone
google-genai
python-dotenv
